backend/requirement.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
Django==3.1.4
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
mysql-connector-python==8.0.22
mysqlclient @ file:///C:/Users/Chanuka/Desktop/Internship%20Tutorials/myProjectInter/mysqlclient-1.4.6-cp39-cp39-win_amd64.whl
protobuf==3.14.0
PyJWT==1.7.1
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
